Move Tabler interface to tablename.go
